Detech and fix float conversions, fix rand_intv() and correct LOG_INGO typos

+++ lib/parser.c
@@ -1914,7 +1914,7 @@ read_timer(vector_t *strvec, size_t index, unsigned long *res, unsigned long min
 	fmax_time = (double)((max_time) ? max_time : TIMER_MAXIMUM) / TIMER_HZ;
 
 	ret = read_double_strvec(strvec, index, &timer, fmin_time, fmax_time, ignore_error);
-	*res = timer * TIMER_HZ > TIMER_MAXIMUM ? TIMER_MAXIMUM : timer * TIMER_HZ;
+	*res = timer * TIMER_HZ > TIMER_MAXIMUM ? TIMER_MAXIMUM : (unsigned long)(timer * TIMER_HZ);
 
 	return ret;
 }
